Change to findFactorial()
Changed data type of return value from int to double to maintain consistency with homework document.

+++ GitDemo.java
@@ -23,7 +23,7 @@ public class GitDemo {
      *  the function will return the value of (1 * 2 * 3 * 4)
      *****/
     public double findFactorial(int num) {
-        int factorialNumber = 1; // change data type to int since factorials can only ever return integers
+        double factorialNumber = 1; // change data type to int since factorials can only ever return integers
         
         for(int i = 1; i <= num; ++i){
             factorialNumber = factorialNumber * i;
